requirements.txt
fastapi_amis_admin==0.7.2
fastapi_user_auth==0.7.3
fastapi-user-auth[jwt]
fastapi-scheduler==0.0.15
fastapi==0.111.1
uvicorn==0.34.0
gunicorn==23.0.0
hypercorn==0.17.3
loguru==0.7.3
simplejson==3.20.1
jinja2==3.1.5
pysondb-v2==2.2.0
pydantic_settings==2.1.0
pydantic[dotenv]==2.3.0
aiomysql==0.2.0
sqlmodelx==0.0.12
sqlmodel==0.0.20
jaeger-client==4.8.0
Tornado>=4.0.0,<5.0.0
opentracing_instrumentation>=3.3.1
Starlette-OpenTracing>=0.1.0
nest_asyncio>=1.6.0
cryptography>=44.0.2
requests>=2.32.3
 emails>=0.6
